create migration for servie tables

+++ LightFireMoreTech5.Data/Entities/AtmService.cs
@@ -3,8 +3,8 @@ using System.ComponentModel.DataAnnotations.Schema;
 namespace LightFireMoreTech5.Data.Entities;
 public class AtmService
 {
-    [ForeignKey("atms")] public int atmId { get; set; }
+    [ForeignKey("atms")] public long atmId { get; set; }
     public Atm Atm { get; set; }
-    [ForeignKey("services")] public int serviceId { get; set; }
+    [ForeignKey("services")] public long serviceId { get; set; }
     public Service Service { get; set; }
 }
