+++ LightFireMoreTech5.Data/Entities/OfficeService.cs
@@ -3,8 +3,8 @@ using System.ComponentModel.DataAnnotations.Schema;
 namespace LightFireMoreTech5.Data.Entities;
 public class OfficeService
 {
-    [ForeignKey("offices")] public int officeId { get; set; }
+    [ForeignKey("offices")] public long officeId { get; set; }
     public Office Office { get; set; }
-    [ForeignKey("services")] public int serviceId { get; set; }
+    [ForeignKey("services")] public long serviceId { get; set; }
     public Service Service { get; set; }
 }
